// src/core/model/caseStudy.ts
export interface CaseStudy {
	id?: number;
	name?: string;
	document?: string;
	industrieId?: string;
	industrieIdName?: string;
	shortDescription?: string;
	isActive?: boolean;
	image?: string;
	createDate?: Date;
	updateDate?: Date;
	deleteDate?: Date;
	createById?: number;
	updateById?: number;
	deleteById?: number;
	isDelete?: boolean;
}